Refactor import statements to use the updated utility module path and improve exception handling in API tests.
All test are run sucessfully
