+++ tests/geometry/fixtures/io.py
@@ -5,7 +5,7 @@ from unittest import mock
 import pytest
 import requests
 
-from owimetadatabase_preprocessor.utils import dict_generator
+from owimetadatabase_preprocessor.utility.utils import dict_generator
 
 
 @pytest.fixture
